Revert "Merge branch 'main' of https://github.com/grafana/grafana into fetch_applied_alert_configs_history"
This reverts commit 892fffc77320b6cdbc46a2352514b8aecae5ebfd, reversing
changes made to 4b8655cd576729ca0d457c1ac9f082ede36ff5ea.

+++ pkg/services/store/entity/sqlstash/querybuilder.go
@@ -5,7 +5,7 @@ import "strings"
 type selectQuery struct {
 	fields   []string // SELECT xyz
 	from     string   // FROM object
-	limit    int64
+	limit    int
 	oneExtra bool
 
 	where []string
